requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.2
jinja2==3.1.4
pyyaml==6.0.2
sqlalchemy==2.0.35
psycopg2-binary
passlib[bcrypt]
bcrypt==4.0.1
pydantic[email]
reportlab
aiohttp
